# frontend/openapi/identity-service.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API APSAS identity-service",
    "description": "Tài liệu API cho APSAS identity-service",
    "version": "1.0"
  },
  "servers": [{ "url": "/", "description": "mặc định" }],
...
  "paths": {
    "/api/v1/users": {
      "get": {
        "tags": ["Quản lý người dùng"],
        "summary": "Lấy tất cả người dùng",
        "description": "Lấy danh sách người dùng có phân trang và sắp xếp (chỉ Admin)",
        "operationId": "getAllUsers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Số trang (bắt đầu từ 0)",
            "required": false,
            "schema": { "type": "string", "minimum": 0 },
            "example": 0
          },
          {
            "name": "size",
            "in": "query",
            "description": "Số lượng phần tử mỗi trang",
            "required": false,
            "schema": { "type": "string", "maximum": 100, "minimum": 1 },
            "example": 10
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PageResponseUserResponse" } } }
          },
          "default": {
            "description": "Phản hồi lỗi mặc định",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } }
          }
        },
        "security": [{ "Bearer Authentication": [] }]
      },
      "post": {
        "tags": ["Quản lý người dùng"],
        "summary": "Tạo người dùng",
        "description": "Tạo mới người dùng (chỉ Admin)",
        "operationId": "createUser",
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateUserRequest" } } },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } } }
          },
          "default": {
            "description": "Phản hồi lỗi mặc định",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } }
          }
        },
        "security": [{ "Bearer Authentication": [] }]
      }
    },
    "/api/v1/users/{userId}/deactivate": {
      "post": {
        "tags": ["Quản lý người dùng"],
        "summary": "Vô hiệu hóa người dùng",
        "description": "Vô hiệu hóa tài khoản người dùng (chỉ Admin)",
        "operationId": "deactivateUser",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "No Content" },
          "default": {
            "description": "Phản hồi lỗi mặc định",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } }
          }
        },
        "security": [{ "Bearer Authentication": [] }]
      }
    },
    "/api/v1/users/{userId}/activate": {
      "post": {
        "tags": ["Quản lý người dùng"],
        "summary": "Kích hoạt người dùng",
        "description": "Kích hoạt tài khoản người dùng (chỉ Admin)",
        "operationId": "activateUser",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "No Content" },
          "default": {
            "description": "Phản hồi lỗi mặc định",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } }
          }
        },
        "security": [{ "Bearer Authentication": [] }]
      }
    },
...
    "/api/auth/verify-email": {
      "post": {
        "tags": ["Xác thực"],
        "summary": "Xác thực email",
        "description": "Xác thực email người dùng bằng mã token",
        "operationId": "verifyEmail",
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TokenRequest" } } },
          "required": true
        },
        "responses": {
          "204": { "description": "No Content" },
          "default": {
            "description": "Phản hồi lỗi mặc định",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } }
          }
        }
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "tags": ["Xác thực"],
        "summary": "Đặt lại mật khẩu",
        "description": "Đặt lại mật khẩu bằng mã token",
        "operationId": "resetPassword",
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResetPasswordRequest" } } },
          "required": true
        },
        "responses": {
          "204": { "description": "No Content" },
          "default": {
            "description": "Phản hồi lỗi mặc định",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } }
          }
        }
      }
    },
    "/api/auth/resend-verification": {
      "post": {
        "tags": ["Xác thực"],
        "summary": "Gửi lại email xác thực",
        "description": "Gửi lại liên kết xác thực email",
        "operationId": "resendVerificationEmail",
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EmailRequest" } } },
          "required": true
        },
        "responses": {
          "204": { "description": "No Content" },
          "default": {
            "description": "Phản hồi lỗi mặc định",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": ["Xác thực"],
        "summary": "Đăng ký người dùng mới",
        "description": "Đăng ký tài khoản mới với vai trò sinh viên",
        "operationId": "register",
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } } },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } } }
          },
          "default": {
            "description": "Phản hồi lỗi mặc định",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Xác thực"],
        "summary": "Đăng nhập",
        "description": "Xác thực người dùng và trả về JWT token",
        "operationId": "login",
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } } }
          },
          "default": {
            "description": "Phản hồi lỗi mặc định",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } }
          }
        }
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "tags": ["Xác thực"],
        "summary": "Yêu cầu đặt lại mật khẩu",
        "description": "Yêu cầu gửi liên kết đặt lại mật khẩu qua email",
        "operationId": "requestPasswordReset",
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EmailRequest" } } },
          "required": true
        },
        "responses": {
          "204": { "description": "No Content" },
          "default": {
            "description": "Phản hồi lỗi mặc định",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } }
          }
        }
      }
    },
...
    "/api/v1/users/{userId}": {
      "get": {
        "tags": ["Quản lý người dùng"],
        "summary": "Lấy người dùng theo ID",
        "description": "Lấy thông tin người dùng theo ID (chỉ Admin và Giảng viên)",
        "operationId": "getUserById",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } } }
          },
          "default": {
            "description": "Phản hồi lỗi mặc định",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } }
          }
        },
        "security": [{ "Bearer Authentication": [] }]
      },
      "delete": {
        "tags": ["Quản lý người dùng"],
        "summary": "Xóa người dùng",
        "description": "Xóa tài khoản người dùng (chỉ Admin)",
        "operationId": "deleteUser",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "No Content" },
          "default": {
            "description": "Phản hồi lỗi mặc định",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } }
          }
        },
        "security": [{ "Bearer Authentication": [] }]
      }
    },
    "/api/v1/users/role/{role}": {
      "get": {
        "tags": ["Quản lý người dùng"],
        "summary": "Lấy người dùng theo vai trò",
        "description": "Lấy danh sách người dùng theo vai trò có phân trang và sắp xếp (chỉ Admin và Giảng viên)",
        "operationId": "getUsersByRole",
        "parameters": [
          {
            "name": "role",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "enum": ["STUDENT", "INSTRUCTOR", "CONTENT_PROVIDER", "ADMIN"] }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Số trang (bắt đầu từ 0)",
            "required": false,
            "schema": { "type": "string", "minimum": 0 },
            "example": 0
          },
          {
            "name": "size",
            "in": "query",
            "description": "Số lượng phần tử mỗi trang",
            "required": false,
            "schema": { "type": "string", "maximum": 100, "minimum": 1 },
            "example": 10
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PageResponseUserResponse" } } }
          },
          "default": {
            "description": "Phản hồi lỗi mặc định",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } }
          }
        },
        "security": [{ "Bearer Authentication": [] }]
      }
    }
  },
  "components": {
    "schemas": {
      "CreateUserRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email", "minLength": 1 },
          "password": { "type": "string", "maxLength": 2147483647, "minLength": 8 },
          "firstName": { "type": "string", "maxLength": 100, "minLength": 0 },
          "lastName": { "type": "string", "maxLength": 100, "minLength": 0 },
          "role": { "type": "string", "enum": ["STUDENT", "INSTRUCTOR", "CONTENT_PROVIDER", "ADMIN"] },
          "isActive": { "type": "boolean" },
          "isEmailVerified": { "type": "boolean" }
        },
        "required": ["email", "firstName", "lastName", "password"]
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "role": { "type": "string", "enum": ["STUDENT", "INSTRUCTOR", "CONTENT_PROVIDER", "ADMIN"] },
          "isActive": { "type": "boolean" },
          "isEmailVerified": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "TokenRequest": {
        "type": "object",
        "properties": { "token": { "type": "string", "minLength": 1 } },
        "required": ["token"]
      },
      "ResetPasswordRequest": {
        "type": "object",
        "properties": {
          "token": { "type": "string", "minLength": 1 },
          "newPassword": { "type": "string", "maxLength": 2147483647, "minLength": 8 }
        },
        "required": ["newPassword", "token"]
      },
      "EmailRequest": {
        "type": "object",
        "properties": { "email": { "type": "string", "format": "email", "minLength": 1 } },
        "required": ["email"]
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email", "minLength": 1 },
          "password": { "type": "string", "maxLength": 2147483647, "minLength": 8 },
          "firstName": { "type": "string", "maxLength": 100, "minLength": 0 },
          "lastName": { "type": "string", "maxLength": 100, "minLength": 0 }
        },
        "required": ["email", "firstName", "lastName", "password"]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "type": { "type": "string" },
          "user": { "$ref": "#/components/schemas/UserResponse" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email", "minLength": 1 },
          "password": { "type": "string", "minLength": 1 }
        },
        "required": ["email", "password"]
      },
...
      "PageResponseUserResponse": {
        "type": "object",
        "description": "Lớp bọc phản hồi phân trang",
        "properties": {
          "content": {
            "type": "array",
            "description": "Danh sách phần tử trong trang hiện tại",
            "items": { "$ref": "#/components/schemas/UserResponse" }
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32",
            "description": "Số trang hiện tại (bắt đầu từ 0)",
            "example": 0
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "description": "Số lượng phần tử mỗi trang",
            "example": 10
          },
          "totalElements": {
            "type": "integer",
            "format": "int64",
            "description": "Tổng số phần tử trên tất cả các trang",
            "example": 100
          },
          "totalPages": { "type": "integer", "format": "int32", "description": "Tổng số trang", "example": 10 },
          "first": { "type": "boolean", "description": "Có phải trang đầu tiên không", "example": true },
          "last": { "type": "boolean", "description": "Có phải trang cuối cùng không", "example": false },
          "hasNext": { "type": "boolean", "description": "Có trang tiếp theo không", "example": true },
          "hasPrevious": { "type": "boolean", "description": "Có trang trước đó không", "example": false }
        }
      },
      "ProblemDetail": {
        "description": "Chi tiết lỗi theo chuẩn RFC 9457",
        "properties": {
          "type": { "format": "uri", "description": "Loại lỗi (URI)" },
          "title": { "description": "Tiêu đề lỗi" },
          "status": { "format": "int32", "description": "Mã trạng thái HTTP" },
          "detail": { "description": "Mô tả chi tiết lỗi" },
          "instance": { "format": "uri", "description": "Đường dẫn instance lỗi" }
        }
      }
    },
    "securitySchemes": { "Bearer Authentication": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } }
  }
}